Restructure folder directory Folder structure has been updated to match  https://github.com/team-bimmer/ribbit/issues/50  (Core Architecture RFC)
