+++ lib/cli/commands/build.js
@@ -1,7 +1,7 @@
 const { exec } = require('child_process');
 const path = require('path');
 const rimraf = require('rimraf');
-const { linkUserDeps, unlinkUserDeps } = require('../utils');
+const { linkUserDeps, unlinkUserDeps } = require('../utils/index');
 
 const ribbitConfig = require(path.join(process.env.PWD, '/ribbit.config.js'));
