+++ lib/cli/commands/init.js
@@ -3,7 +3,7 @@ const path = require('path');
 const fs = require('fs');
 const util = require('util');
 const chalk = require('chalk');
-const { prettifyJS } = require('../utils');
+const { prettifyJS } = require('../utils/index');
 
 const promptQuestions = require('../prompts/prompts_init');
