+++ lib/cli/prompts/prompts_init.js
