+++ lib/cli/utils/index.js
