+++ lib/cli/utils/prettifyJS.js
