+++ lib/getRibbitStore.js
