+++ lib/index.js
@@ -1,7 +1,7 @@
-const build = require('./commands/build.js');
-const dependencies = require('./commands/dependencies.js');
-const init = require('./commands/init');
-const workspaces = require('./commands/workspaces');
+const build = require('./cli/commands/build.js');
+const dependencies = require('./cli/commands/dependencies.js');
+const init = require('./cli/commands/init');
+const workspaces = require('./cli/commands/workspaces');
 
 module.exports = {
   build,
