+++ lib/ribbitPreload.js
